requirements.txt
# Core
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9

# Filtering & search
django-filter>=24.0

# CORS
django-cors-headers>=4.3

# Async task queue (structure only)
celery>=5.4
redis>=5.0

# Configuration
python-decouple>=3.8

# WebSocket / real-time
channels[daphne]>=4.1
channels-redis>=4.0

# Testing
pytest>=8.0
pytest-django>=4.8
